Fix: Node closes its second implementation audit - public form security, installer and 6.3 update, integrity of comments and points, the document tree, moderate, display modes, code rules and plan documents of stages S20.1-S20.8
The second audit of stages S00-S19.8 split its findings into eight fix windows. This commit carries all of them:
every finding was re-checked in code before its fix, the forks the contracts left open were decided by the user
and written into the owning documents, and the plan of docs/node is complete.

Core changes:

1. Public form security - S20.1 (modules/node/index.php, core/classes/node/query.php, core/classes/node/service.php):
- A guest passes the comment captcha on every POST before any file is read; uploads of the public form need
  checkEditorUploadAccess() and respect the role max and maxfiles, inactive and link roles take no file
- A new per-IP window limits.send (60 seconds) answers NodeException::LIMITED with 429, the moderator of the
  type is exempt; a closed category answers 404 through NodeQuery::checkNodeCategory()
- Search titles of Node are decoded once and escaped once

2. Installer and 6.3 update - S20.2 (setup/index.php, core/classes/pdo.php, setup/lang, UPGRADING.md):
- The form never prints the database password, an empty field keeps the stored one, a refusal keeps
  config/setup.unlock
- The table prefix must match [A-Za-z0-9_]{1,32} and the admin file name goes through filterVar()
- setUpdateModules() keeps the switches a 6.2 site stored in _modules over the shipped config/modules.php;
  checkUpdateBase() refuses a clean install over existing tables of the prefix and an update without
  _users and _admins

3. Integrity of comments, points and categories - S20.3 (core/classes/comment.php, core/classes/node/service.php, admin/modules/categories.php):
- deleteNode() refuses with STORAGE when Point cannot confirm the compensation of an award
- A premoderated comment of a Node material locks the material; NodeService::checkTypeRegistry() guards the
  category screen, which now answers _ERROR_UP instead of a raw SQL failure

4. Document tree - S20.4 (modules/node/index.php, templates/lite, setup/sql):
- A type with features.tree shows the tree of its documents with the current node, previous and next links
  and a paged remainder, read in batches of NodeQuery::TREEPART
- _nodes gets the index tree (tid, status, id) in table.sql and table_update6_3.sql

5. The moderate action and small rendering fixes - S20.5 (core/classes/node/service.php, core/classes/comment.php, core/classes/node/view.php):
- Approving a material, a comment or an asset report writes a Point event moderate with the aid of the
  moderator; comment event scopes live in one place
- NodeView renders only card for a NodeTarget, the unused fragment node/search.html is removed, the sitemap
  reads Node in batches of limits.syncbatch, related cards no longer count views

6. Display modes - S20.6 (modules/node/index.php, templates/lite, tools/ui-contract.php):
- docs, faq, files and media get their own card and view templates, media also its list
- Cards and og:image prefer the poster role, the list card carries download, every view gets
  download_label and hits_label

7. Code rules and consolidation - S20.7 (core/system.php, modules/node, core/classes, admin/modules, setup/index.php, tools/node-profile.php):
- getNodeReader(), getNodeWriter() and addNodeMail() live in the core next to getNodeHandler(); support and
  admin mail go through addNodeMail()
- render_blocks() becomes setBlockView(int $bid); arrow functions and parameters of the plan code are typed,
  redundant casts are removed

8. Plan documents - S20.8 (docs/node):
- Status lines, the test structure of 13 and 02, the NodeContext and target signatures of 05, the cover
  label of 06 and the module tree of 04 follow the code

Benefits:
- Every finding of the second audit is fixed or recorded with its reason in docs/node/PROGRESS.md
- Promised but missing features of the plan - the document tree, moderate and display modes - are shipped

Technical notes:
- Schema: KEY tree (tid, status, id) on _nodes in table.sql and table_update6_3.sql
- Config: config/node.php gets limits.send = 60; NodeException gains LIMITED = 6
- API: render_blocks() is replaced by setBlockView(int $bid); getConst() resolves only names with a leading
  underscore; checkUpdateBase() takes bool $fresh; NodeService::checkTypeRegistry() is new; getNodeReader(),
  getNodeWriter() and addNodeMail() are core functions
- Breaking: the installer refuses a prefix outside [A-Za-z0-9_]{1,32}, so a 6.2 site with such a prefix cannot
  pass the update; the theme fragment fragments/node/search.html is gone; NodeView accepts only card for targets

+++ admin/modules/scheduler.php
@@ -283,7 +283,7 @@ function save(): void {
         }
         $sett = $issys ? ((isset($curr['settings']) && is_array($curr['settings'])) ? $curr['settings'] : []) : ['url' => $url];
         if ($issys && isset(SCHED_LIMITS[$name], $sett['limit'])) {
-            $lim = (int)getVar('post', 'limit', 'num', 0);
+            $lim = getVar('post', 'limit', 'num', 0);
             if ($lim < 1 || $lim > SCHED_LIMITS[$name]) {
                 setRedirect($afile.'.php?name=scheduler&op=add&job='.$name);
                 return;
